Extract duplicated code.
